+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- function `queryWithin`
+
 ## [1.1.2] (2021-04-02)
 
 This version only updates the README with function descriptions.