+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] (2021-04-03)
+
 ### Added
 
 - function `queryWithin`
@@ -77,7 +79,8 @@ This is merely a documentation update.
 
 [docs]: https://jjwesterkamp.github.io/lambda-dom/
 
-[Unreleased]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.1.2...HEAD
+[Unreleased]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.1.2...v1.2.0
 [1.1.2]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/JJWesterkamp/lambda-dom/compare/v1.0.1...v1.1.0